requirements.txt
Flask
pytest
pytest-cov
python-dotenv
flake8
celery
gunicorn
kubernetes
flask-login
bcrypt
requests
flask-marshmallow
marshmallow-sqlalchemy
psycopg2
Flask-SQLAlchemy
WTForms
Flask-WTF
Flask-Mail
Flask-Caching
Flask-Limiter
flask-cors
